changed date format + fixed trimming
(1) I put the date format in yyyy-mm-dd instead of the default
which is mm-dd-yyyy in order to make sorting easier

(2) I initally thought that some columns had extra whitespace
only at the beginning and not the end, but I was wrong. So I added
trimming from the end as well.